Refactor: move engine::Command to standalone file
